requirements.txt
Django==4.1
djangorestframework==3.13.1
firebase-admin==5.2.0
drf-spectacular>=0.15.1,<0.16
django-rest-swagger==2.1.2
pyrebase
PyJWT